# perf(python): speedup datetime conversion